Separate the requests logic and the friends logic from the messenger component

+++ src/app/components/navbar/navbar.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { AuthService } from "../../services/auth.service";
 import { WebSocketService } from "../../socket/WebSocketService";
 import { Router } from "@angular/router";
-import { MessageService } from "../../services/message.service";
+import { MessageService } from "../../shared/message.service";
 import { MessageType } from "../../models/enums/MessageType";
 import {SimpleNotif} from "../../models/SimpleNotif";
 import {NotificationType} from "../../models/enums/NotificationType";
@@ -23,43 +23,45 @@ export class NavbarComponent implements OnInit {
   }
 
   ngOnInit(): void {
+    if(this.authService.isLoggedIn()) {
+      this.friendsRequestService.getReceivedUnseenRequests().subscribe({
+        next: (requests) => {
+          this.messageService.setUnseenRequest(
+            requests.map((req) => ({
+              senderId: req.sender.connectionId,
+              senderUsername: req.sender.connectionUsername,
+              receiverId: req.receiver.connectionId,
+              receiverUsername: req.receiver.connectionUsername,
+              notificationType: NotificationType.REQUEST
+            }))
+          );
+        },
+        error: (err) => console.log(err),
+      });
 
-    this.friendsRequestService.getReceivedUnseenRequests().subscribe({
-      next: (requests) => {
-        this.messageService.setUnseenRequest(
-          requests.map((req) => ({
-            senderId: req.sender.connectionId,
-            senderUsername: req.sender.connectionUsername,
-            receiverId: req.receiver.connectionId,
-            receiverUsername: req.receiver.connectionUsername,
+      this.messageService.message$.subscribe(message => {
+        if (message) {
+
+          const simpleNotif: SimpleNotif = {
+            senderId: message.senderId,
+            senderUsername: message.senderUsername,
+            receiverId: message.receiverId,
+            receiverUsername: message.receiverUsername,
             notificationType: NotificationType.REQUEST
-          }))
-        );
-      },
-      error: (err) => console.log(err),
-    });
+          }
 
-    this.messageService.message$.subscribe(message => {
-      if (message) {
+          if (message.messageType === MessageType.FRIEND_REQUEST) {
+            this.messageService.addUnseenRequest(simpleNotif);
+          }
 
-        const simpleNotif: SimpleNotif = {
-          senderId: message.senderId,
-          senderUsername: message.senderUsername,
-          receiverId: message.receiverId,
-          receiverUsername: message.receiverUsername,
-          notificationType: NotificationType.REQUEST
+          if (message.messageType === MessageType.FRIEND_REQUEST_CANCELED) {
+            this.messageService.removeUnseenRequest(message.senderId, message.receiverId);
+          }
         }
 
-        if (message.messageType === MessageType.FRIEND_REQUEST) {
-          this.messageService.addUnseenRequest(simpleNotif);
-        }
+      });
+    }
 
-        if (message.messageType === MessageType.FRIEND_REQUEST_CANCELED) {
-          this.messageService.removeUnseenRequest(message.senderId, message.receiverId);
-        }
-      }
-
-    });
   }
 
   handleLogout() {
